requirements.txt
celery[redis]==4.4.2
ddt==1.3.1
django==3.0.4
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.8
django-filter==2.2.0
django-oauth-toolkit==1.3.0
django-redis==4.11.0
djangorestframework==3.11.0
factory-boy==2.12.0
methodtools==0.1.2
psycopg2==2.8.4
uwsgi==2.0.18
